web/requirements.txt
Django==2.2.17
smartmin==2.3.0
psycopg2-binary
django-choices
django_chartit